Verify pdf generation works and is properly sent to the client.
We shouldn't really test the content of the PDF here. We'd essentially
be testing Prawn, and while view specs are something we can do we don't
normally go through the effort.

Here we're just verifying that the PDF can be generated without error
and with the correct status code and headers.