Pull in old png Solidus logo
The Solidus admin area was recently changed to use an svg logo
https://github.com/solidusio/solidus/pull/2995
but that is not supported by Prawn so it broke the pdf embed here.

The old png logo is no longer used in Solidus so it's only a matter of
time before someone removes it. As such I have pulled it in for use
in the pdf header, but kept the same name because I figured it less
likely that the png logo ever be updated again in Solidus than maybe
someone customized their Solidus install by overwriting the image
instead of changing the config.

This has the unfortunate affect of no longer automatically following
Solidus admin logo customizations, but I don't think it would be worth
doing fancy svg detection at that point. We have our own config, folks
can use that if needed.